Refactor admin server actions
