+++ nerin-electric-site-v3-fixed/app/admin/constants.ts
@@ -1,0 +1,26 @@
+export type PackFormState = {
+  success: boolean
+  error?: string
+}
+
+export const initialPackFormState: PackFormState = {
+  success: false,
+}
+
+export type MaintenanceFormState = {
+  success: boolean
+  error?: string
+}
+
+export const initialMaintenanceFormState: MaintenanceFormState = {
+  success: false,
+}
+
+export type CaseStudyFormState = {
+  success: boolean
+  error?: string
+}
+
+export const initialCaseStudyState: CaseStudyFormState = {
+  success: false,
+}
